requirements.txt
accelerate==1.0.1
beautifulsoup4==4.12.3
deepspeed==0.16.3
flax==0.10.2
ftfy==6.3.1
huggingface_hub==0.25.2
importlib_metadata==8.5.0
invisible_watermark==0.2.0
jax==0.5.0
k_diffusion==0.1.1.post1
librosa==0.10.2.post1
lpips==0.1.4
matplotlib==3.7.5
msgpack_python==0.5.6
note_seq==0.0.5
numba==0.58.1
numpy==1.24.4
omegaconf==2.1.1
onnxruntime==1.20.1
opencv_python==4.1.2.30
opencv_python_headless==4.10.0.84
packaging==24.2
paintmind==0.0
Pillow==10.0.1
Pillow==11.1.0
piqa==1.3.2
Requests==2.32.3
safetensors==0.5.2
scikit_learn==1.3.2
scipy==1.9.1
torch==2.4.1
torchsde==0.2.6
torchvision==0.19.1
tqdm==4.66.6
transformers==4.46.3
xformers==0.0.29.post1
